refactor(frontend): extract reusable list docs service
